third_party/blink/web_tests/platform/generic/external/wpt/secure-payment-confirmation/enrollment-in-iframe.sub.https-expected.txt

This is a testharness.js-based test.
FAIL SPC enrollment in cross-origin iframe assert_equals: expected null but got object "NotSupportedError: The user agent does not support public key credentials."
PASS SPC enrollment in cross-origin iframe fails without user activation
PASS SPC enrollment in cross-origin iframe without payment permission
Harness: the test ran to completion.
